@@ -1,26 +1,38 @@
 {
     "name": "lawtext",
-    "displayName": "Lawtext",
+    "version": "0.0.6",
     "description": "Lawtext syntax highlight and preview",
-    "version": "0.0.5",
-    "publisher": "yamachi",
-    "engines": {
-        "vscode": "^1.18.0"
-    },
-    "license": "SEE LICENSE IN LICENSE",
     "homepage": "https://github.com/yamachig/vscode-lawtext/blob/master/README.md",
     "repository": {
         "type": "git",
         "url": "https://github.com/yamachig/vscode-lawtext.git"
     },
-    "categories": [
-        "Languages"
-    ],
+    "license": "SEE LICENSE IN LICENSE",
+    "main": "./out/extension",
+    "scripts": {
+        "compile": "tsc -p ./",
+        "postinstall": "node ./node_modules/vscode/bin/install",
+        "vscode:prepublish": "npm run compile",
+        "watch": "tsc -watch -p ./"
+    },
+    "dependencies": {
+        "lawtext": "^0.0.3"
+    },
+    "devDependencies": {
+        "@types/node": "^9.4.6",
+        "typescript": "^2.8.3",
+        "vscode": "^1.1.10"
+    },
+    "engines": {
+        "vscode": "^1.23.0"
+    },
     "activationEvents": [
         "onCommand:extension.showLawtextPreview",
         "onLanguage:lawtext"
     ],
-    "main": "./out/extension",
+    "categories": [
+        "Programming Languages"
+    ],
     "contributes": {
         "languages": [
             {
@@ -75,20 +87,6 @@
             ]
         }
     },
-    "scripts": {
-        "vscode:prepublish": "npm run compile",
-        "compile": "tsc -p ./",
-        "watch": "tsc -watch -p ./",
-        "postinstall": "node ./node_modules/vscode/bin/install"
-    },
-    "devDependencies": {
-        "@types/mocha": "^2.2.44",
-        "@types/node": "^8.0.58",
-        "@types/nunjucks": "^3.0.0",
-        "typescript": "^2.6.2",
-        "vscode": "^1.1.10"
-    },
-    "dependencies": {
-        "nunjucks": "^3.0.1"
-    }
+    "displayName": "Lawtext",
+    "publisher": "yamachi"
 }
